+++ src/main/java/org/moqucu/games/nightstalker/model/GameWorld.java
@@ -8,7 +8,7 @@ import java.util.*;
 public class GameWorld {
 
     @Getter
-    private long time = 0;
+    private double time = 0;
 
     @Getter
     private final Map<String, GameObject> objects = new HashMap<>();
@@ -19,7 +19,7 @@ public class GameWorld {
     @Getter
     private final MazeGraphV2 mazeGraph = new MazeGraphV2();
 
-    public void pulse(long milliseconds) {
+    public void pulse(double milliseconds) {
 
         time += milliseconds;
         timeListeners.forEach(timeListener -> timeListener.elapseTime(milliseconds));
